Ignore spaceship operator lint check
Ensuring correct ordering of foreman::plugin resources is more important than
the error of realising virtual foreman::plugin resources, so ignoring this lint
for now.

If a tag was added to our use of foreman::plugin then other users of this
defined type would always have to add it too, which isn't user friendly.