Skip defined() check for Foreman URL in CLI class on Puppet 3.6/7
Puppet 3.7.4 and below are not able to check whether a variable is
defined in another class due to bug PUP-4072, so foreman::foreman_url
isn't used as defined() is returning false on these versions.

This change skips the defined() check for these versions to ensure it
always picks the URL from foreman::foreman_url, at the expense of not
supporting strict variables before Puppet 3.7.5.