Fixes #18757 - Handle dynflow service in foreman core
The dynflow service was moved into core with 1.17. That renamed it to
dynflowd and means the service should be started by default.

For compatibility parameters are provided. This means we must move the
service name detection to the jobs.pp file because that's the only place
we know the actual value.