Refactor Puppet handling
This moves all puppetmaster defaults to its own parameters file.
Parameters that were used in both are duplicated. It also adds Puppet
types to all parameters.

This removes puppet settings from the top level. The puppet_ssldir is
now an internal detail to determine defaults. It is also no longer set
in Foreman since it's only used as a base for determining SSL paths
which we explicitly set. There is a PR open to remove this setting
altogether.