Fixes #32827 - Add sendmail config options
As part of CVE-2021-3584 the option email_sendmail_location was limited
to just 4 choices. This allows admins to set it via settings.yaml. The
idea is that if you can edit settings.yaml, you're already compromised
while UI could be less protected.

When a setting is present in settings.yaml, the option becomes read-only
in the UI.

If the options are not set, they don't show up in settings.yaml.