Drop server_ssl_certs_dir parameter
This parameter could be used to configure a directory with CA
certificates to be trusted. This parameter is generally not used and
$server_ssl_cert is preferred.

The parameter was introduced in 08911c3a6c776462fcc1aa99103fe588b8feb365
to override the default in puppetlabs-apache. Prior to version 2.1.0 it
used to default to the system store, which created a security problem.
In puppetlabs-apache 2.1.0 the default was changed to undef[1]. Since
this module depends on >= 5.5.0, it's safe to assume we don't need it.

It can still be set via Hiera:

    foreman::config::apache::https_vhost_options:
      ssl_certs_dir: /path/to/certs/dir

[1]: https://github.com/puppetlabs/puppetlabs-apache/commit/7bb35c2293c12ce52329a4391fe1f20389efef06