Fixes #34161 - Run apipie:cache:index after db:migrate
With Puppet you can write:

    Class['A'] ~> Class['B'] ~> Class['C']

This implies Class['A'] ~> Class['C'] and rspec-puppet actually will
tell you that chaining is there, but it doesn't actually happen. This
was reported in https://github.com/rodjek/rspec-puppet/pull/821.

In this particular case it means that if DB seeding doesn't happen then
the apipie caches indexes aren't refreshed. By chaining it to db:migrate
there's a much bigger chance it actually happens.

Normally apipie:cache:index needs to run after a package is updated so
ideally this would actually be done in packaging, but this is the
workaround we've been using for a long time.