Fixes #36582 - Detect logging layout based on type
When the file logger is used the multiline_request_pattern layout should
be used by default, but with journald the request ID is already stored
as structured data with the log, so the pattern layout is better. The
the best layout for the logging type is now used if no layout is
specified.